Restore void* DobbyHook call style and silence build warnings
- Add DobbyHook void*/void** compat shim so call sites keep original
  API style against LSPosed's typed dobby_dummy_func_t signature
- KittyMemory read2HexStr: use std::vector instead of VLAs
- Dobby: guard LINE_MAX redefinition, fix snprintf truncation size,
  widen arm_shift_c r_val to uint64_t (fixes shift-overflow warning
  and latent asr sign-extension truncation)

+++ Includes/Dobby/source/InstructionRelocation/arm/InstructionRelocationARM.cc
@@ -96,7 +96,7 @@ uint32_t arm_shift_c(uint32_t val, uint32_t shift_type, uint32_t shift_count, ui
                      uint32_t *carry_out) {
     if (shift_count == 0)
         return val;
-    uint32_t r_val;
+    uint64_t r_val;
     uint32_t carry = carry_in;
     switch (shift_type) {
         case arm_shift_lsl:
